+++ libraries/PixelShared/Sprite.h
@@ -49,7 +49,7 @@ public:
             
             
         }
-        for(int i=0;i<children.size();i++)
+        for(size_t i=0;i<children.size();i++)
         {
             children[i]->int_update();
         }
